Update table link to link to placement request
